requirements.txt
Flask==2.3.2
pdfplumber==0.9.0
pymupdf==1.22.5
pytesseract==0.3.10
Pillow==10.0.0
spacy==3.6.1